+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-## Master
-* Fixed a connection leak in `_send_bulk()`.
+## Version 3.11.3 (2026-08-29)
+* Fixed a connection leak in `_send_bulk()`. Thanks @simonphilips!
 
 ## Version 3.11.2 (2026-04-16)
 * `_send_email` should dispatch email using it's own connection. Thanks @ibadarrohman!